add cmd control, change dictionary size, file functions don't work

+++ LZ77/Algorithms/ArrayExtension.cs
@@ -13,7 +13,7 @@ namespace LZ77.Algorithms
     public static class ArrayExtension
     {
         /// <summary>
-        /// Shift elements in array
+        /// Shift elements in array, delete <paramref name="offset"/> elements from left/right side
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="array">array</param>
@@ -21,7 +21,7 @@ namespace LZ77.Algorithms
         /// <param name="direction">direction: left | right</param>
         /// <param name="elementsCount">how many elements are in array (to improve efficiency)</param>
         /// <returns>returns new array</returns>
-        public static T[] ShiftElements<T>(T[] array, int offset, ushort elementsCount, ShiftDirection direction)
+        public static T[] ShiftElements<T>(T[] array, int offset, int elementsCount, ShiftDirection direction)
         {
             T[] arr = new T[array.Length];
             var cnt = Math.Min(elementsCount, array.Length - offset);
